Add getEntriesWithChildFrames() method to Performance
This change introduces a method to include the Performance Entries of nested, same-origin iframes when fetching Performance Timeline data. It uses a Breadth-First Search algorithm on the Frame Tree whose root node is the current frame.

This functionality is available via the `includeFrames` boolean flag in each of `getEntries()`, `getEntriesByType()`, and `getEntriesByName()`.

Change-Id: I725d29c8295851d957c8b6580bd0d065f3bb197f