[bfcache] Reset NotRestoredReasons for redirect
This CL implements resetting of NotRestoredReasons when redirect happens.
For example, a.com->b.com->(back but server redirect)c.com
When this happens, we should not be reporting the reasons for a.com for c.com.

This CL does this by adding NotRestoredReasons as a part of commit navigation params instead of a parameter in CommitNavigation() method.

This is to implement what we discussed here:
https://chromium-review.googlesource.com/c/chromium/src/+/3986625/5..12/content/browser/renderer_host/back_forward_cache_impl.cc#1551

Change-Id: Idee05c88d1073fbe5c3d19c0b9edcd360ce6c1a5